Move revcache implementation to lib
So that it can also be used by the future go PS
